requirements.txt
python-telegram-bot==21.10
feedparser==6.0.11
beautifulsoup4==4.12.3
requests==2.32.3
python-dotenv==1.0.1
openai>=1.40.0
